refactor gensim by variable, bug-compatible
this version produces identical output to the previous but is
organized with all assignments for a given TRD variable in one place,
rather than all settings for a DH or vent system.  I hope that this
will make it easier to understand the relationship between systems.
It makes adding a neXw system marginally more tedious, but debugging
easier.  It is already easy to see what is set in any single system,
by inspecting csv or TRD.

In rewriting this code, I uncovered a number of bugs in the old
gensim, as well as several intended but unmade changes.  These should
be in the following commits.