Import gzip'd dat files in hourly_data
This was a test to see if decompressing would be faster than
reading larger files from disk.  On my laptop (X120e), it is
slower (2.6 s vs 3.6 sec).  I'm keeping the code in case the decreased
storage space is useful.

gzip is better than zip because python has better support.  The compression
rates are similar.  Zip files with many scenarios are useful for collaboration
because zip is better supported on Windows.  But for analysis, it's better to
have each scenario in its own directory.

This is all orthogonal to the optimization by combining the several dat files
into one per scenario.